requirements.txt
requests
huggingface_hub
bitnet
llama-cpp-python
sentencepiece
transformers
pytest
flake8
pydantic